+++ src/main/kotlin/com/waffletoy/team1server/post/controller/PostController.kt
@@ -2,7 +2,7 @@ package com.waffletoy.team1server.post.controller
 
 import com.waffletoy.team1server.post.dto.*
 import com.waffletoy.team1server.post.service.PostService
-import com.waffletoy.team1server.post.service.S3Service
+import com.waffletoy.team1server.s3.service.S3Service
 import com.waffletoy.team1server.user.AuthUser
 import com.waffletoy.team1server.user.AuthUserOrNull
 import com.waffletoy.team1server.user.dtos.User
@@ -20,8 +20,6 @@ import org.springframework.web.bind.annotation.*
 @Validated
 class PostController(
     private val postService: PostService,
-    private val s3Service: S3Service,
-    @Value("\${amazon.aws.bucket}") private val bucketName: String,
 ) {
     // 채용 공고 상세 페이지 불러오기
     @GetMapping("/{post_id}")
@@ -101,24 +99,7 @@ class PostController(
         )
     }
 
-    // s3 bucket
-    @PostMapping("/upload/presigned")
-    fun generateUploadPresignedUrl(
-        @Parameter(hidden = true) @AuthUser user: User,
-        @RequestBody preSignedUploadReq: PreSignedUploadReq,
-    ): ResponseEntity<PresignedURL> {
-        val presignedUrl = s3Service.generateUploadPreSignUrl(user, preSignedUploadReq, bucketName)
-        return ResponseEntity.ok(PresignedURL(presignedUrl))
-    }
 
-    @PostMapping("/download/presigned")
-    fun generateDownloadPresignedUrl(
-        @Parameter(hidden = true) @AuthUser user: User,
-        @RequestBody preSignedDownloadReq: PreSignedDownloadReq,
-    ): ResponseEntity<PresignedURL> {
-        val presignedUrl = s3Service.generateDownloadPreSignUrl(user, preSignedDownloadReq, bucketName)
-        return ResponseEntity.ok(PresignedURL(presignedUrl))
-    }
 
     // dev
     @PostMapping("/dev/make-dummy")
@@ -239,19 +220,6 @@ data class Paginator(
 data class PostWithPage(
     val posts: List<PostBrief>,
     val paginator: Paginator,
-)
-
-data class PreSignedUploadReq(
-    val fileName: String,
-    val fileType: String,
-)
-
-data class PreSignedDownloadReq(
-    val fileName: String,
-)
-
-data class PresignedURL(
-    val presignedUrl: String,
 )
 
 data class PasswordRequest(val pw: String)
